Added code for initial population.
All of the code is in src/.  src/config.lisp is just the inputs for a sample
run.  Also, I added information so that it can be compiled with scons.  It
requires a shell script called combine to be located in ~/bin.

The current code takes command-line arguments --initial-pop, the number of
candidates in the population, and --initial-depth, the (rough) maximum depth
of any tree in the population.  It doesn't do any mutations, crossover, or
generation activity at all.